requirements.txt
pydub
requests
pyannote
azure-cognitiveservices-speech